08903: Refactor moved burn package inside hts package
